SEC SaaS Metrics5/backend/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
boto3==1.21.22
botocore==1.24.22
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
fake-useragent==0.1.11
feedparser==6.0.8
idna==3.3
importlib-metadata==4.11.2
jmespath==1.0.0
lxml==4.8.0
Markdown==3.3.6
multitasking==0.0.10
numpy==1.22.3
pandas==1.4.1
parse==1.19.0
psycopg2==2.9.3
pyee==8.2.2
pymongo==3.12.1
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-html==0.10.0
s3transfer==0.5.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.2.4
tqdm==4.63.0
urllib3==1.26.8
w3lib==1.22.0
websockets==10.2
yahoo-fin==0.8.9.1
yfinance==0.1.70
zipp==3.7.0